Add files via upload
Food Donation Project is a mission to end hunger and no wasting of food to make a hungry-free world. According to the latest survey, 1.3 billion tons of food is thrown as waste every year. Additionally, one-third of the food consumed is stated as leftovers. The focus of this project is to reduce the amount of food wasted and being used to the needy people. Therefore, an android-based application is developed by which a person can donate food with their capacity and at the same time the application lets the organization to put their request on their requirements. The project contains modules in which the admin adds the details such as name of the restaurant, location of the restaurant, food details such as type of food and amount of food serves for people and the user can view all these details in the user login to know about the food details and interact with them. Languages used in the front end is JAVA and the back end is firebase.